requirements.txt
asgiref==3.5.2
Django==4.1
django-debug-toolbar==3.6.0
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Pillow==9.2.0
PyJWT==2.4.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
